+++ crates/gui/src/covers.rs
@@ -5,13 +5,18 @@ use crate::{Logic, data_dir::DATA_DIR};
 use anyhow::Result;
 use slint::Weak;
 use std::fs;
-use twbm_core::{covers::download_cover, game_id::GameID};
+use twbm_core::{config::PreferredLanguage, covers::download_cover, game_id::GameID};
 
-pub fn download_covers(ids: Vec<GameID>, weak: Weak<Logic<'static>>) -> Result<()> {
-    fs::create_dir_all(DATA_DIR.join("covers"))?;
+pub fn download_covers(
+    ids: Vec<GameID>,
+    preferred_language: PreferredLanguage,
+    weak: &Weak<Logic<'static>>,
+) -> Result<()> {
+    let covers_dir = DATA_DIR.join("covers");
+    fs::create_dir_all(&covers_dir)?;
 
     for (i, game_id) in ids.into_iter().enumerate() {
-        if download_cover(game_id, &DATA_DIR).is_ok() {
+        if download_cover(game_id, &covers_dir, preferred_language).is_ok() {
             let _ = weak.upgrade_in_event_loop(move |logic| {
                 logic.invoke_reload_cover(i as i32);
             });